+++ src/selectors/analyticsSelectors.js
@@ -347,3 +347,8 @@ export const selectChartDatapointsCPM = createSelector(
 		return result
 	}
 )
+
+export const selectPublisherRevenueNoticeActive = createSelector(
+	selectPublisherTotalImpressions,
+	totalImpressions => !totalImpressions
+)