+++ src/selectors/tablesDataSelectors.js
@@ -26,12 +26,16 @@ export const selectCampaignsTableData = createSelector(
 
 			const firstUnit = adUnits[0] || {}
 
+			const to = `/dashboard/${side}/campaigns/${id}`
+			const toReceipt = `/dashboard/${side}/receipt/${id}`
+
 			return {
 				media: {
-					side: side,
-					id: item.id,
+					side,
+					id,
 					mediaUrl: firstUnit.mediaUrl || '',
 					mediaMime: firstUnit.mediaMime || '',
+					to,
 				},
 				title: item.title,
 				status: {
@@ -40,12 +44,12 @@ export const selectCampaignsTableData = createSelector(
 				},
 				depositAmount: Number(formatUnits(item.depositAmount || '0', decimals)),
 				fundsDistributedRatio: item.status.fundsDistributedRatio || 0,
-				impressions: selectCampaignEventsCount('IMPRESSION', item.id),
-				clicks: selectCampaignEventsCount('CLICK', item.id),
+				impressions: selectCampaignEventsCount('IMPRESSION', id),
+				clicks: selectCampaignEventsCount('CLICK', id),
 
 				ctr:
-					(selectCampaignEventsCount('CLICK', item.id) /
-						selectCampaignEventsCount('IMPRESSION', item.id)) *
+					(selectCampaignEventsCount('CLICK', id) /
+						selectCampaignEventsCount('IMPRESSION', id)) *
 						100 || 0,
 
 				minPerImpression: {
@@ -63,9 +67,11 @@ export const selectCampaignsTableData = createSelector(
 				withdrawPeriodStart:
 					spec.withdrawPeriodStart || item.withdrawPeriodStart,
 				actions: {
-					side: side,
+					side,
 
 					id,
+					to,
+					toReceipt,
 					receiptReady:
 						(item.status.humanFriendlyName === 'Closed' ||
 							item.status.humanFriendlyName === 'Completed') &&
@@ -114,17 +120,19 @@ export const selectAdSlotsTableData = createSelector(
 	(slots, side) =>
 		slots.map(item => {
 			const id = item.id || item.ipfs
+			const to = `/dashboard/${side}/slots/${id}`
 			return {
 				media: {
 					id,
 					mediaUrl: item.fallbackUnit ? `ipfs://${item.fallbackUnit}` : '', //TODO: provide fallback image to slot
 					mediaMime: 'image/jpeg',
+					to,
 				},
 				title: item.title,
 				type: item.type.replace('legacy_', ''),
 				created: item.created,
 				actions: {
-					to: `/dashboard/${side}/AdSlot/${id}`,
+					to,
 					item,
 				},
 			}
@@ -133,11 +141,13 @@ export const selectAdSlotsTableData = createSelector(
 
 export const selectAdUnitsTableData = createSelector(
 	[
-		(state, { side, campaignId }) => ({
+		(state, { side, campaignId, items }) => ({
 			side,
-			items: campaignId
-				? selectCampaignUnitsById(state, campaignId)
-				: selectAdUnits(state),
+			items:
+				items ||
+				(campaignId
+					? selectCampaignUnitsById(state, campaignId)
+					: selectAdUnits(state)),
 			impressionsByAdUnit: id =>
 				campaignId
 					? selectCampaignAnalyticsByChannelToAdUnit(state, {
@@ -163,12 +173,14 @@ export const selectAdUnitsTableData = createSelector(
 	({ side, items, impressionsByAdUnit, clicksByAdUnit }) =>
 		Object.values(items).map(item => {
 			const id = item.id || item.ipfs
+			const to = `/dashboard/${side}/units/${id}`
 			return {
 				id,
 				media: {
 					id,
 					mediaUrl: item.mediaUrl,
 					mediaMime: item.mediaMime,
+					to,
 				},
 				impressions: impressionsByAdUnit(id) || 0,
 				clicks: clicksByAdUnit(id) || 0,
@@ -177,7 +189,7 @@ export const selectAdUnitsTableData = createSelector(
 				type: item.type,
 				created: item.created,
 				actions: {
-					to: `/dashboard/${side}/AdUnit/${id}`,
+					to,
 					item,
 				},
 			}